Well, this kills me, but here we are.
It turns out that a known problem with using Flasks decorators + multiple files
is... you pretty much are guaranteed a circular import. I spent as much time as
I could stand trying to hack around this, and did not succeed. So, here we are.
Circular again.
